refactor: extract history mapper for readability
